+++ common/src/main/java/traben/entity_model_features/EMFClient.java
@@ -10,6 +10,8 @@ public class EMFClient {
 
     public static final String MOD_ID = "entity_model_features";
 
+    public static final long START_TIME = System.currentTimeMillis();
+
     //@Override
     public static void init() {
         LogManager.getLogger().info("[Entity Model Features]: Loading! 1.19.3");